Single menu item version split
